requirements.txt
streamlit
speechrecognition
pyttsx3
python-dotenv
google-generativeai
pyaudio
setuptools
gtts